feat: implemented rename and new folder actions
