+++ src/components/tabs/tab-new.tsx
@@ -17,7 +17,7 @@ export default () => {
     const handleOpenFile = useOpenFile(async (document, file) => {
         const drive = documentDrive?.state.drives[0]; // TODO improve default drive selection
         if (drive) {
-            const node = await addFile(file, file.name, drive.id);
+            const node = await addFile(file, drive.id, file.name);
             if (node) {
                 openFile(drive.id, node.path, selectedTab);
             }
